Move Asset input components into their own files
