+++ src/components/Form/AudioInput.js
@@ -1,0 +1,25 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import FileInput from './FileInput';
+import { Audio } from '../Assets';
+
+const AudioInput = ({ value, onChange }) => (
+  <FileInput
+    value={value}
+    onChange={onChange}
+    accept="audio/*"
+  >
+    { url => (<Audio url={url} controls />) }
+  </FileInput>
+);
+
+AudioInput.propTypes = {
+  value: PropTypes.string,
+  onChange: PropTypes.func.isRequired,
+};
+
+AudioInput.defaultProps = {
+  value: '',
+};
+
+export default AudioInput;
