+++ src/components/Form/index.js
@@ -4,3 +4,6 @@ export { default as TextInput } from './TextInput';
 export { default as OptionsInput } from './OptionsInput';
 export { default as MarkdownInput } from './MarkdownInput';
 export { default as FileInput } from './FileInput';
+export { default as ImageInput } from './ImageInput';
+export { default as AudioInput } from './AudioInput';
+export { default as VideoInput } from './VideoInput';
