+++ src/components/StageEditor/sections/ContentItem.js
@@ -1,40 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { MarkdownInput, FileInput } from '../../../components/Form';
-import { Image, Audio, Video } from '../../../components/Assets';
-
-// eslint-disable-next-line
-const ImageInput = ({ value, onChange }) => (
-  <FileInput
-    value={value}
-    onChange={onChange}
-    accept="image/*"
-  >
-    { url => (<Image url={url} alt="preview" />) }
-  </FileInput>
-);
-
-// eslint-disable-next-line
-const AudioInput = ({ value, onChange }) => (
-  <FileInput
-    value={value}
-    onChange={onChange}
-    accept="audio/*"
-  >
-    { url => (<Audio url={url} autoplay controls />) }
-  </FileInput>
-);
-
-// eslint-disable-next-line
-const VideoInput = ({ value, onChange }) => (
-  <FileInput
-    value={value}
-    onChange={onChange}
-    accept="video/*"
-  >
-    { url => (<Video controls url={url} />) }
-  </FileInput>
-);
+import { MarkdownInput, ImageInput, AudioInput, VideoInput } from '../../../components/Form';
 
 const contentInputs = {
   text: MarkdownInput,
